# cloudify/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cloudify API
  description: |
    Cloudify is a self-hosted deployment platform for modern web applications.
    This API provides programmatic access to manage projects, deployments, domains,
    environment variables, serverless functions, and more.

    ## Authentication
    All API requests require authentication using a Bearer token.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-api-token>
    ```

    ## Rate Limiting
    API requests are rate limited to 1000 requests per minute per API key.

    ## Errors
    The API uses standard HTTP response codes to indicate success or failure.
    - 2xx: Success
    - 4xx: Client error (bad request, unauthorized, not found)
    - 5xx: Server error
  version: 1.0.0
  contact:
    name: Cloudify Support
    email: support@cloudify.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://cloudify.tranthachnguyen.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List all projects
      description: Returns a list of all projects accessible to the authenticated user.
      operationId: listProjects
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
          description: Search by project name
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Projects]
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, repoUrl]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                repoUrl:
                  type: string
                  format: uri
                repoBranch:
                  type: string
                  default: main
                framework:
                  type: string
                  enum: [nextjs, react, vue, nuxt, svelte, astro, static]
                buildCommand:
                  type: string
                outputDirectory:
                  type: string
                installCommand:
                  type: string
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /projects/{projectId}/env:
    get:
      tags: [Environment Variables]
      summary: List environment variables
      operationId: listEnvVars
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: environment
          in: query
          schema:
            type: string
            enum: [production, preview, development]
      responses:
        '200':
          description: List of environment variables
          content:
            application/json:
              schema:
                type: object
                properties:
                  variables:
                    type: array
                    items:
                      $ref: '#/components/schemas/EnvVariable'

    post:
      tags: [Environment Variables]
      summary: Create environment variable
      operationId: createEnvVar
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, value]
              properties:
                key:
                  type: string
                value:
                  type: string
                environment:
                  type: string
                  enum: [production, preview, development]
                  default: production
                isSecret:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Variable created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvVariable'
...
  /webhooks/bitbucket:
    post:
      tags: [Webhooks]
      summary: Bitbucket webhook endpoint
      operationId: bitbucketWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
      description: Project ID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        repoUrl:
          type: string
        repoBranch:
          type: string
        framework:
          type: string
        buildCommand:
          type: string
        outputDirectory:
          type: string
        installCommand:
          type: string
        status:
          type: string
          enum: [active, paused, deleted]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    EnvVariable:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        value:
          type: string
          description: Masked for secrets
        environment:
          type: string
          enum: [production, preview, development]
        isSecret:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []